Update to flywheel-sdk 12.0.0
- modify manifest to remove optional configs so this can be run
from the SDK.
- add release-drafter
