+++ utils/license/__init__.py
@@ -1,0 +1,9 @@
+#!/usr/bin/env python3
+"""Software license handling utilities.
+
+Software license files can be provided to the platform in various
+ways.  Each module in this package handles the license for a particular
+software package.
+"""
+
+# vi:set autoindent ts=4 sw=4 expandtab : See Vim, :help 'modeline'
